Issue #96 closed by applying recommend fix, static void setClockDivider(BYTE); changed to static void setClockDivider(WORD);
Former-commit-id: e7ed2ea872b65b9ec238e7eeb2efe1fc070e2f60

+++ hardware/pic32/libraries/SPI/SPI.h
@@ -92,7 +92,7 @@ public:
 
   static void setBitOrder(BYTE);
   static void setDataMode(WORD);
-  static void setClockDivider(BYTE);
+  static void setClockDivider(WORD);
 };
 
 extern SPIClass SPI;
